fixed test

+++ src/modules/person/persistence/person.repository.integration-spec.ts
@@ -48,7 +48,6 @@ describe('PersonRepository', () => {
     let usernameGeneratorService: DeepMocked<UsernameGeneratorService>;
     let personPermissionsMock: DeepMocked<PersonPermissions>;
     let configService: ConfigService;
-    let personRepositoryMock: DeepMocked<PersonRepository>;
 
     beforeAll(async () => {
         module = await Test.createTestingModule({
@@ -82,8 +81,6 @@ describe('PersonRepository', () => {
         kcUserServiceMock = module.get(KeycloakUserService);
         usernameGeneratorService = module.get(UsernameGeneratorService);
         configService = module.get(ConfigService);
-
-        personRepositoryMock = createMock<PersonRepository>();
 
         await DatabaseTestModule.setupDatabase(orm);
     }, DEFAULT_TIMEOUT_FOR_TESTCONTAINERS);
@@ -647,15 +644,31 @@ describe('PersonRepository', () => {
                 personPermissionsMock.getOrgIdsWithSystemrecht.mockResolvedValueOnce([person1.id]);
 
                 await em.persistAndFlush(mapper.map(person1, PersonDo, PersonEntity));
-
-                personRepositoryMock.checkIfDeleteIsAllowed.mockResolvedValueOnce({
+                await sut.getPersonIfAllowed(person1.id, personPermissionsMock);
+                const personGetAllowed: Result<Person<true>> = await sut.getPersonIfAllowed(
+                    person1.id,
+                    personPermissionsMock,
+                );
+                if (!personGetAllowed.ok) {
+                    throw new EntityNotFoundError('Person', person1.id);
+                }
+                const checkIfDeleteIsAllowedSpy: jest.SpyInstance<
+                    Promise<Result<Person<true>, Error>>,
+                    [personId: string, permissions: PersonPermissions]
+                > = jest.spyOn(sut, 'checkIfDeleteIsAllowed').mockResolvedValue({
                     ok: false,
                     error: new EntityCouldNotBeDeleted('Person', person1.id),
                 });
 
+                await sut.checkIfDeleteIsAllowed(personGetAllowed.value.id, personPermissionsMock);
+
                 const result: Result<void, DomainError> = await sut.deletePerson(person1.id, personPermissionsMock);
 
                 expect(result.ok).toBeFalsy();
+                if (!result.ok) {
+                    expect(result.error).toBeInstanceOf(EntityCouldNotBeDeleted);
+                }
+                checkIfDeleteIsAllowedSpy.mockRestore();
             });
             it('should not delete the person because it has no keycloakId', async () => {
                 const person1: PersonDo<true> = DoFactory.createPerson(true);
